Refactor to Remove isinstance Calls
Refactored out most of the isinstance calls various methods in
deck.models.

You might note that isinstance still exists in Pile.push. Given that
the only appropriate use of discard involves either a Card of a list of Cards,
and that we use the Pile interface, and not the Card interface to discard cards,
isinstance makes a lot of sense here.